Updated AliasTab stats widget
