--- lib/screens/alias_tab/components/components_export.dart
@@ -1,5 +1,0 @@
-export 'alias_animated_list.dart';
-export 'alias_shimmer_loading.dart';
-export 'alias_tab_pie_chart.dart';
-export 'aliases_stats_shimmer.dart';
-export 'empty_list_alias_tab.dart';
